requirements.txt
# requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2
python-pptx==0.6.21
openpyxl==3.1.2